First pass at refactoring the code base into multiple class instead of one big class
