+++ python-client/src/main/python/janusgraph_grpc_python/graph_operation/graph_adder.py
@@ -1,0 +1,17 @@
+
+
+class GraphElementAdder:
+    SINGLE = False
+    ALL = False
+    name = None
+
+    def __init__(self):
+        pass
+
+    def set(self, retriever):
+        if retriever == "ALL":
+            self.ALL = True
+        else:
+            self.SINGLE = True
+            self.name = retriever
+        return self
